feat: 流水线插件支持展示获得的荣誉和SLA等信息 #8114

+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/common/impl/StoreTotalStatisticServiceImpl.kt
@@ -353,7 +353,7 @@ class StoreTotalStatisticServiceImpl @Autowired constructor(
         )
         var value = 0.0
         result.forEach {
-            value += it.value1() as Double
+            value += it.value1() as Int
         }
         return if (pluralFlag) value / 2 else value
     }
